MxShop/requirements.txt
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.19
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-formtools==2.2
django-guardian==2.3.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-reversion==3.0.9
djangorestframework==3.12.2
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.19.0
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.3
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
